Expand changelog line and move test to adapter zone
