+++ .changes/unreleased/Fixes-20231031-144837.yaml
@@ -1,5 +1,5 @@
 kind: Fixes
-body: Fix exception running empty seed file
+body: Fix compilation exception running empty seed file and support new Integer agate data_type
 time: 2023-10-31T14:48:37.774871-04:00
 custom:
   Author: gshank
