+++ tests/adapter/dbt/tests/adapter/simple_seed/test_seed.py
@@ -365,3 +365,26 @@ class TestSeedSpecificFormats(SeedConfigBase):
     def test_simple_seed(self, project):
         results = run_dbt(["seed"])
         assert len(results) == 3
+
+
+class BaseTestEmptySeed:
+    @pytest.fixture(scope="class")
+    def project_config_update(self):
+        return {
+            "seeds": {
+                "quote_columns": False,
+            },
+        }
+
+    @pytest.fixture(scope="class")
+    def seeds(self):
+        return {"empty_with_header.csv": "a,b,c"}
+
+    def test_empty_seeds(self, project):
+        # Should create an empty table and not fail
+        results = run_dbt(["seed"])
+        assert len(results) == 1
+
+
+class TestEmptySeed(BaseTestEmptySeed):
+    pass
